Renamed classes and split them up
