+++ Assets/LeapMotionModules/InteractionEngine/Scripts/InteractionBehaviours/InteractionGraspEvents.cs
@@ -7,13 +7,8 @@ namespace Leap.Unity.Interaction {
   [RequireComponent(typeof(InteractionBehaviourBase))]
   public class InteractionBehaviourEvents : MonoBehaviour {
 
-    [Serializable]
-    public class HandEvent : UnityEvent<Hand> { }
-
     public UnityEvent onGraspBegin;
     public UnityEvent onGraspEnd;
-    public HandEvent onHandGrasp;
-    public HandEvent onHandRelease;
 
     private InteractionBehaviourBase _interactionBehaviour;
 
@@ -23,8 +18,6 @@ namespace Leap.Unity.Interaction {
       if (_interactionBehaviour != null) {
         _interactionBehaviour.OnGraspBeginEvent += onGraspBegin.Invoke;
         _interactionBehaviour.OnGraspEndEvent += onGraspEnd.Invoke;
-        _interactionBehaviour.OnHandGraspedEvent += onHandGrasp.Invoke;
-        _interactionBehaviour.OnHandReleasedEvent += onHandRelease.Invoke;
       }
     }
 
@@ -32,8 +25,6 @@ namespace Leap.Unity.Interaction {
       if (_interactionBehaviour != null) {
         _interactionBehaviour.OnGraspBeginEvent -= onGraspBegin.Invoke;
         _interactionBehaviour.OnGraspEndEvent -= onGraspEnd.Invoke;
-        _interactionBehaviour.OnHandGraspedEvent -= onHandGrasp.Invoke;
-        _interactionBehaviour.OnHandReleasedEvent -= onHandRelease.Invoke;
       }
     }
   }
