+++ Assets/LeapMotionModules/InteractionEngine/Scripts/InteractionBehaviours/InteractionHandEvents.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0dca9b92bb693af4c8cef5df04ae6805
+timeCreated: 1471027665
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
